+++ docs/pages/getting-started/changelog.md
@@ -5,7 +5,7 @@ description: Getting started with Sloth.css, see what's new.
 
 ## Changelog
 
-Sloth.css uses [semantic versioning](https://semver.org/) to indicate breaking changes, major changes with downward compatibility and minor patches.
+Sloth.css uses [semantic versioning](https://semver.org/) to indicate breaking changes, major changes with downward compatibility or minor changes.
 
 ### 0.1.0