+++ docs/pages/getting-started/changelog.md
@@ -8,6 +8,26 @@ tags: release version changes
 
 Sloth.css uses [semantic versioning](https://semver.org/) to indicate breaking changes, major changes with downward compatibility or minor changes.
 
-### 0.1.0
+### 0.2.0-beta
+
+- 📜 Add documentation ([#1](https://github.com/devmount/sloth.css/pull/1), [#32](https://github.com/devmount/sloth.css/pull/32), [#38](https://github.com/devmount/sloth.css/pull/38))
+- 🚀 Build tool ([#2](https://github.com/devmount/sloth.css/pull/2), [#15](https://github.com/devmount/sloth.css/pull/15))
+- 📜 Code of conduct ([#7](https://github.com/devmount/sloth.css/pull/7))
+- 📜 Organize repository ([#8](https://github.com/devmount/sloth.css/pull/8), [#9](https://github.com/devmount/sloth.css/pull/9), [#10](https://github.com/devmount/sloth.css/pull/10), [#13](https://github.com/devmount/sloth.css/pull/13))
+- 📜 Contribution guidelines ([#11](https://github.com/devmount/sloth.css/pull/11))
+- 🔨 Fix deprecated overlow class ([#14](https://github.com/devmount/sloth.css/pull/14))
+- ➕ Component: Badge ([#16](https://github.com/devmount/sloth.css/pull/16))
+- ➕ Component: Separator ([#18](https://github.com/devmount/sloth.css/pull/18))
+- ➕ Component: Callout ([#25](https://github.com/devmount/sloth.css/pull/25))
+- ➕ Component: Breadcrumb ([#26](https://github.com/devmount/sloth.css/pull/26))
+- ➕ Component: State ([#27](https://github.com/devmount/sloth.css/pull/27))
+- ➕ Component: Toast ([#28](https://github.com/devmount/sloth.css/pull/28))
+- 💚 Color improvements ([#29](https://github.com/devmount/sloth.css/pull/29))
+- ➕ Component: Button ([#30](https://github.com/devmount/sloth.css/pull/30), [#34](https://github.com/devmount/sloth.css/pull/34))
+- ➕ Component: Loader ([#31](https://github.com/devmount/sloth.css/pull/31))
+- ⚡️ Performance optimizations ([#33](https://github.com/devmount/sloth.css/pull/33))
+- ➕ Input invalid state ([#39](https://github.com/devmount/sloth.css/pull/39))
+
+### 0.1.0-beta
 
 - 🚀 Initial release